requirements-dev.txt
arrow==1.2.3 ; python_version >= "3.10" and python_version < "4.0"
atomicwrites==1.4.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
attrs==22.2.0 ; python_version >= "3.10" and python_version < "4.0"
binaryornot==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
black==22.12.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
chardet==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cookiecutter==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
distro==1.8.0 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.9.0 ; python_version >= "3.10" and python_version < "4.0"
flake8-black==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
flake8-builtins==1.5.3 ; python_version >= "3.10" and python_version < "4.0"
flake8-comprehensions==3.10.1 ; python_version >= "3.10" and python_version < "4.0"
flake8-debugger==4.1.2 ; python_version >= "3.10" and python_version < "4.0"
flake8-fixme==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
flake8-variables-names==0.0.5 ; python_version >= "3.10" and python_version < "4.0"
flake8==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
identify==2.5.17 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
isort==5.12.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2-time==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
mccabe==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
nodeenv==1.7.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.11.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==2.6.2 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pre-commit==2.21.0 ; python_version >= "3.10" and python_version < "4.0"
py==1.11.0 ; python_version >= "3.10" and python_version < "4.0"
pycodestyle==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
pyflakes==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
pytest-cookies==0.6.1 ; python_version >= "3.10" and python_version < "4.0"
pytest==6.2.5 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-slugify==8.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
ruamel-yaml-clib==0.2.7 ; platform_python_implementation == "CPython" and python_version < "3.12" and python_version >= "3.10"
ruamel-yaml==0.17.32 ; python_version >= "3.10" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_full_version < "3.11.0a7"
types-setuptools==57.4.18 ; python_version >= "3.10" and python_version < "4.0"
types-toml==0.10.8.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
virtualenv==20.17.1 ; python_version >= "3.10" and python_version < "4.0"
